+++ Changelog.md
@@ -1,13 +1,13 @@
 # Todo list of React Module Project.
 
-### Updated to commit #20 | Date 4/04/2025:
+### Updated to commit #21 | Date 4/04/2025:
 
 ## Changes:
 
 ### Bugs Fixes:
 
--   Overflow set to hidden to all elements in [Error.page.css](./website/src/styles/Error.page.css) instead to all child elements of elements with the class `error-page-wrapper`.
--   Fixed the warning error of not using a key in rendering the paragraph lines in [Header.tsx - p element](./website/src/components/Header.tsx). Changed it to place a `span` element with a `br` element inside it.
+-   Added [`SearchBox`](./website/src/components/SearchBox.tsx) component to the navbar.
+-   Finished all tasks up this point except the login functionality.
 
 ## Todo: